feat(shapeshift): remove dead code
