+++ packages/blockchain-wallet-v4-frontend/src/data/components/exchange/exchange.sagas.ts
@@ -830,11 +830,13 @@ export default ({ api, coreSagas, networks }) => {
             actions.alerts.displaySuccess(C.FIRST_PAX_TRADE_INFO, null, true)
           )
           yield put(
+            // @ts-ignore
             actions.modals.showModal(RESULTS_MODAL, formatExchangeTrade(trade))
           )
         }
       } else {
         yield put(
+          // @ts-ignore
           actions.modals.showModal(RESULTS_MODAL, formatExchangeTrade(trade))
         )
       }
