--- packages/blockchain-wallet-v4-frontend/src/modals/Exchange/ShapeshiftTradeDetails/selectors.js
@@ -1,35 +1,0 @@
-import { getCoinFromPair } from 'services/ShapeshiftService'
-import { selectors } from 'data'
-
-export const getData = (depositAddress, state) => {
-  const tradeR = selectors.core.kvStore.shapeShift.getTrade(
-    depositAddress,
-    state
-  )
-
-  const transform = trade => {
-    const { status, quote } = trade
-    const {
-      pair,
-      quotedRate,
-      minerFee,
-      orderId,
-      depositAmount,
-      withdrawalAmount
-    } = quote
-    const { sourceCoin, targetCoin } = getCoinFromPair(pair)
-
-    return {
-      sourceCoin,
-      targetCoin,
-      status,
-      quotedRate,
-      minerFee,
-      orderId,
-      depositAmount,
-      withdrawalAmount
-    }
-  }
-
-  return tradeR.map(transform)
-}
