+++ packages/blockchain-wallet-v4-frontend/src/modals/types.ts
@@ -1,7 +1,7 @@
 export type ModalPropsType = {
   close: () => void
-  closeAll: () => void
+  closeAll?: () => void
   position: number
   total: number
-  userClickedOutside: boolean
+  userClickedOutside?: boolean
 }
