--- packages/blockchain-wallet-v4/src/network/api/shapeShift/index.js
@@ -1,14 +1,0 @@
-export default ({ get }) => {
-  const url = 'https://shapeshift.io/'
-
-  const getTradeStatus = address =>
-    get({
-      url,
-      endPoint: `txStat/${address}`,
-      contentType: 'application/json'
-    })
-
-  return {
-    getTradeStatus
-  }
-}
