+++ packages/blockchain-wallet-v4/src/redux/kvStore/actionTypes.js
@@ -5,7 +5,6 @@ import * as contacts from './contacts/actionTypes'
 import * as eth from './eth/actionTypes'
 import * as lockbox from './lockbox/actionTypes'
 import * as root from './root/actionTypes'
-import * as shapeShift from './shapeShift/actionTypes'
 import * as userCredentials from './userCredentials/actionTypes'
 import * as whatsNew from './whatsNew/actionTypes'
 import * as xlm from './xlm/actionTypes'
@@ -14,7 +13,6 @@ export {
   root,
   whatsNew,
   eth,
-  shapeShift,
   buySell,
   contacts,
   bch,
