+++ packages/blockchain-wallet-v4/src/redux/kvStore/actions.js
@@ -5,7 +5,6 @@ import * as contacts from './contacts/actions'
 import * as eth from './eth/actions'
 import * as lockbox from './lockbox/actions'
 import * as root from './root/actions'
-import * as shapeShift from './shapeShift/actions'
 import * as userCredentials from './userCredentials/actions'
 import * as whatsNew from './whatsNew/actions'
 import * as xlm from './xlm/actions'
@@ -14,7 +13,6 @@ export {
   root,
   whatsNew,
   eth,
-  shapeShift,
   buySell,
   contacts,
   bch,
