+++ packages/blockchain-wallet-v4/src/redux/kvStore/middleware.js
@@ -29,7 +29,6 @@ const kvStoreMiddleware = ({
         T.buySell.FETCH_METADATA_BUYSELL_SUCCESS,
         T.contacts.FETCH_METADATA_CONTACTS_SUCCESS,
         T.eth.FETCH_METADATA_ETH_SUCCESS,
-        T.shapeShift.FETCH_METADATA_SHAPESHIFT_SUCCESS,
         T.bch.FETCH_METADATA_BCH_SUCCESS,
         T.btc.FETCH_METADATA_BTC_SUCCESS,
         T.lockbox.FETCH_METADATA_LOCKBOX_SUCCESS,
@@ -43,7 +42,6 @@ const kvStoreMiddleware = ({
         [C.BUYSELL]: A.buySell.fetchMetadataBuySellSuccess,
         [C.CONTACTS]: A.contacts.fetchMetadataContactsSuccess,
         [C.ETH]: A.eth.fetchMetadataEthSuccess,
-        [C.SHAPESHIFT]: A.shapeShift.fetchMetadataShapeshiftSuccess,
         [C.BCH]: A.bch.fetchMetadataBchSuccess,
         [C.BTC]: A.btc.fetchMetadataBtcSuccess,
         [C.LOCKBOX]: A.lockbox.fetchMetadataLockboxSuccess,
