+++ packages/blockchain-wallet-v4/src/redux/kvStore/sagaRegister.js
@@ -5,7 +5,6 @@ import buySell from './buySell/sagaRegister'
 import contacts from './contacts/sagaRegister'
 import eth from './eth/sagaRegister'
 import lockbox from './lockbox/sagaRegister'
-import shapeShift from './shapeShift/sagaRegister'
 import userCredentials from './userCredentials/sagaRegister'
 import whatsNew from './whatsNew/sagaRegister'
 import xlm from './xlm/sagaRegister'
@@ -17,7 +16,6 @@ export default ({ api, networks }) =>
     yield fork(bch({ api, networks }))
     yield fork(btc({ api, networks }))
     yield fork(xlm({ api, networks }))
-    yield fork(shapeShift({ api, networks }))
     yield fork(buySell({ api, networks }))
     yield fork(contacts({ api, networks }))
     yield fork(lockbox({ api, networks }))
