+++ packages/blockchain-wallet-v4/src/redux/kvStore/selectors.js
@@ -5,7 +5,6 @@ import * as contacts from './contacts/selectors'
 import * as eth from './eth/selectors'
 import * as lockbox from './lockbox/selectors'
 import * as root from './root/selectors'
-import * as shapeShift from './shapeShift/selectors'
 import * as userCredentials from './userCredentials/selectors'
 import * as whatsNew from './whatsNew/selectors'
 import * as xlm from './xlm/selectors'
@@ -14,7 +13,6 @@ export {
   root,
   whatsNew,
   eth,
-  shapeShift,
   buySell,
   contacts,
   bch,
