--- packages/blockchain-wallet-v4/src/redux/kvStore/shapeShift/actionTypes.js
@@ -1,29 +1,0 @@
-export const CREATE_METADATA_SHAPESHIFT = '@CORE.CREATE_METADATA_SHAPESHIFT'
-export const ADD_STATE_METADATA_SHAPESHIFT =
-  '@CORE.ADD_STATE_METADATA_SHAPESHIFT'
-
-// ADD_TRADE_METADATA_SHAPESHIFT
-export const ADD_TRADE_METADATA_SHAPESHIFT =
-  '@CORE.ADD_TRADE_METADATA_SHAPESHIFT'
-
-// UPDATE_TRADE_METADATA_SHAPESHIFT
-export const UPDATE_TRADE_METADATA_SHAPESHIFT =
-  '@CORE.UPDATE_TRADE_METADATA_SHAPESHIFT'
-
-// FETCH_METADATA_SHAPESHIFT
-export const FETCH_METADATA_SHAPESHIFT = '@CORE.FETCH_METADATA_SHAPESHIFT'
-export const FETCH_METADATA_SHAPESHIFT_LOADING =
-  '@CORE.FETCH_METADATA_SHAPESHIFT_LOADING'
-export const FETCH_METADATA_SHAPESHIFT_SUCCESS =
-  '@CORE.FETCH_METADATA_SHAPESHIFT_SUCCESS'
-export const FETCH_METADATA_SHAPESHIFT_FAILURE =
-  '@CORE.FETCH_METADATA_SHAPESHIFT_FAILURE'
-
-// FETCH_METADATA_SHAPESHIFT
-export const FETCH_SHAPESHIFT_TRADE = '@CORE.FETCH_SHAPESHIFT_TRADE'
-export const FETCH_SHAPESHIFT_TRADE_LOADING =
-  '@CORE.FETCH_SHAPESHIFT_TRADE_LOADING'
-export const FETCH_SHAPESHIFT_TRADE_SUCCESS =
-  '@CORE.FETCH_SHAPESHIFT_TRADE_SUCCESS'
-export const FETCH_SHAPESHIFT_TRADE_FAILURE =
-  '@CORE.FETCH_SHAPESHIFT_TRADE_FAILURE'
