--- packages/blockchain-wallet-v4/src/redux/kvStore/shapeShift/sagaRegister.js
@@ -1,18 +1,0 @@
-import * as AT from './actionTypes'
-import { takeEvery, takeLatest } from 'redux-saga/effects'
-import sagas from './sagas'
-
-export default ({ api, networks }) => {
-  const kvStoreShapeshiftSagas = sagas({ api, networks })
-
-  return function * coreKvStoreShapeshiftSaga () {
-    yield takeEvery(
-      AT.FETCH_SHAPESHIFT_TRADE,
-      kvStoreShapeshiftSagas.fetchShapeshiftTrade
-    )
-    yield takeLatest(
-      AT.FETCH_METADATA_SHAPESHIFT,
-      kvStoreShapeshiftSagas.fetchMetadataShapeshift
-    )
-  }
-}
